add queryToUrl helper
